configure.ac: Fix default behavior of AC_ARG_WITH(glut) if glut isn't available
Fixes a regression introduced in
406248811eb0dfabf75ae9495b54529ec59cce66

It wrongly sets glut_enabled=yes if glut isn't available and neither
option --with-glut nor --without-glut was given.

The default behavior in that case should be if glut is available then
enable glut else it should disable glut.

To fix this the default value of glut_enabled is set back to yes and in
case --without-glut was given glut_enabled is set to no.

Cc: Ross Burton <ross.burton@intel.com>
Signed-off-by: Andreas Boll <andreas.boll.dev@gmail.com>