wgl: Rudimentary wglUseFontBitmaps sample.
It uses SYSTEM_FONT which actually creates some challenges when
emulating wglUseFontBitmaps:  in spite what
https://msdn.microsoft.com/en-us/library/windows/desktop/dd374392.aspx
implies, GetGlyphOutline(GGO_BITMAP) does not seem to work with certain
fonts.  The only solution is to draw the font charactors with a HBITMAP
like the old Mesa fxwgl.c code used to do.  That too, seems to be the
way that opengl32.dll implements wglUseFontBitmaps.

Reviewed-by: Brian Paul <brianp@vmware.com>