wglinfo: query/print sRGB and float pixel format info
Print the columns in the same place as glxinfo.
Use 3 digit hexadecimal field for printing pixel format number.
Remove the 'caveat' field which is non-existant in WGL.

This also fixes a minor bug when we query WGL_PIXEL_TYPE_ARB.
We previously stored the result in info->pfd.iPixelType which is a BYTE
type but the WGL_TYPE_* values are all larger than 255.  Use a temporary
variable instead.

Reviewed-by: Neha Bhende <bhenden@vmware.com>